Update variable styling in app
TODO: update f4.ioc and f7.ioc to change the queue type to "UartCmd_t" from "UARTcmd_t"

Former-commit-id: 913b329d6d2e0386ed21cf385c1b1d0a090a2a4c [formerly ba0ebf3d63564a37fdae357ffd75b78e569ee101]
Former-commit-id: 386cde692e3259a63fe8220903fe156b478491fa

+++ Common/app/Communication.h
@@ -33,7 +33,7 @@ namespace comm{
 // Constant Declarations
 // ----------------------------------------------------------------------------
 /** Buffer offset at which MPU6050 data is inserted */
-constexpr uint8_t ROBOT_STATE_MPU_DATA_OFFSET = 48;
+constexpr size_t ROBOT_STATE_MPU_DATA_OFFSET = 48;
 
 
 // Types & enums
